[golang] Add example implementation.
This gives a base source to update when starting a kata. As I used this
code initially, I lost time creating my source, figuring out the package
to use (I got taken aback by having tests in the main package, as I
always use a _test package).
So with this change, the users directly get a runnable tests with a
source ready to be updated, and 2 examples of assertions (True, and
Equal).