[golang] Reduce go version.
While the current go version is (at the time of writing) 1.22, not
everybody's machines is running on it. And to use go mod, one has to
have a version at least as high as in the go.mod file.
I suggest then to reduce the minimum version to 1.17 (which is the
minimum version of testify, the only dependency of this code.

The version of Go has however no impact on running the tests, only to
use go.mod, which is maybe unlikely in the context of a Kata, but would
be needed whenever a dependency has to be added.

I do not know how this commit would affect the Renovate Bot, in case it
tries to update minimum required version as soon as possible.